feat: move getPlural/Singular to HasMetadata, Pluralize to model-core
