+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/CustomResource.java
@@ -25,15 +25,11 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.Namespaced;
 import io.fabric8.kubernetes.api.model.ObjectMeta;
-import io.fabric8.kubernetes.client.utils.Pluralize;
 import io.fabric8.kubernetes.model.Scope;
 import io.fabric8.kubernetes.model.annotation.Group;
-import io.fabric8.kubernetes.model.annotation.Plural;
 import io.fabric8.kubernetes.model.annotation.ShortNames;
-import io.fabric8.kubernetes.model.annotation.Singular;
 import io.fabric8.kubernetes.model.annotation.Version;
 import io.sundr.builder.annotations.Buildable;
-import java.util.Locale;
 import java.util.Optional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -46,8 +42,8 @@ import org.slf4j.LoggerFactory;
  * <ul>
  *   <li>group is set using {@link HasMetadata#getGroup(Class)}</li>
  *   <li>version is set using {@link HasMetadata#getVersion(Class)}</li>
- *   <li>singular is set using {@link CustomResource#getSingular(Class)}</li>
- *   <li>plural is set using {@link CustomResource#getPlural(Class)}</li>
+ *   <li>singular is set using {@link HasMetadata#getSingular(Class)}</li>
+ *   <li>plural is set using {@link HasMetadata#getPlural(Class)}</li>
  *   <li>computed CRD name using {@link CustomResource#getCRDName(Class)}</li>
  * </ul>
  *
@@ -98,8 +94,8 @@ public abstract class CustomResource<S, T> implements HasMetadata {
     this.apiVersion = version;
     this.kind = HasMetadata.super.getKind();
     scope = this instanceof Namespaced ? Scope.NAMESPACED.value() : Scope.CLUSTER.value();
-    this.singular = getSingular(clazz);
-    this.plural = getPlural(clazz);
+    this.singular = HasMetadata.super.getSingular();
+    this.plural = HasMetadata.super.getPlural();
     this.crdName = getCRDName(clazz);
     this.served = getServed(clazz);
     this.storage = getStorage(clazz);
@@ -176,15 +172,10 @@ public abstract class CustomResource<S, T> implements HasMetadata {
   }
 
   /**
-   * Retrieves the plural form associated with the specified CustomResource if annotated with {@link Plural} or computes a default value
-   * using the value returned by {@link #getSingular(Class)} as input to {@link Pluralize#toPlural(String)}.
-   *
-   * @param clazz the CustomResource whose plural form we want to retrieve
-   * @return the plural form defined by the {@link Plural} annotation or a computed default value
+   * @deprecated use {@link HasMetadata#getPlural(Class)} instead
    */
   public static String getPlural(Class<?> clazz) {
-    final Plural fromAnnotation = clazz.getAnnotation(Plural.class);
-    return (fromAnnotation != null ? fromAnnotation.value().toLowerCase(Locale.ROOT) : Pluralize.toPlural(getSingular(clazz)));
+    return HasMetadata.getPlural(clazz);
   }
 
   @JsonIgnore
@@ -193,16 +184,10 @@ public abstract class CustomResource<S, T> implements HasMetadata {
   }
 
   /**
-   * Retrieves the singular form associated with the specified CustomResource as defined by the {@link Singular} annotation or
-   * computes a default value (lower-cased version of the value returned by {@link HasMetadata#getKind(Class)}) if the annotation
-   * is not present.
-   *
-   * @param clazz the CustomResource whose singular form we want to retrieve
-   * @return the singular form defined by the {@link Singular} annotation or a computed default value
+   * @deprecated use {@link HasMetadata#getSingular(Class)} instead
    */
   public static String getSingular(Class<?> clazz) {
-    final Singular fromAnnotation = clazz.getAnnotation(Singular.class);
-    return (fromAnnotation != null ? fromAnnotation.value() : HasMetadata.getKind(clazz)).toLowerCase(Locale.ROOT);
+    return HasMetadata.getSingular(clazz);
   }
 
   @JsonIgnore
@@ -218,7 +203,7 @@ public abstract class CustomResource<S, T> implements HasMetadata {
    * @return the CRD name associated with the CustomResource
    */
   public static String getCRDName(Class<?> clazz) {
-    return getPlural(clazz) + "." + HasMetadata.getGroup(clazz);
+    return HasMetadata.getPlural(clazz) + "." + HasMetadata.getGroup(clazz);
   }
 
   @JsonIgnore
